new folders, moving files, editing unit 02
I have created the folders: "images" "audios" and "videos" inside the folder assets to order our visual material acording to the kind of file and not by unit.
The files .gif of Psyche are inside images due to the fact the most of them were images. Only one of them was an animation that I put inside "videos" folder.
I have edited 02.md so it matches with our format and to correct the new path for all the images.
I also wrote a summary for this unit based ond the summary of the end that Psyche wrote as a conclusion for the unit.