Refactor of the CSS / SASS code and redesign of the layouts.
Most of the SASS directives have been removed with the exception
of some global variables defined in _base.scss.

The .scss files have been refactored so that some global and
top level definitions are made in _base.scss, and then there
is one file per layout, and one file for the header and footer.

All media queries have been removed for the time being, as well
as any seo code.